+++ src/components/Post/index.tsx
@@ -1,25 +1,13 @@
-import {
-    Card,
-    CardActions,
-    CardContent,
-    CardHeader,
-    CardMedia,
-    Collapse,
-} from '@material-ui/core'
+import { Card, CardContent, CardHeader, CardMedia } from '@material-ui/core'
 import Avatar from '@material-ui/core/Avatar'
 import { red } from '@material-ui/core/colors'
 import IconButton from '@material-ui/core/IconButton'
 import { createStyles, makeStyles, Theme } from '@material-ui/core/styles'
 import Typography from '@material-ui/core/Typography'
-import EditIcon from '@material-ui/icons/Edit'
-import EventNoteIcon from '@material-ui/icons/EventNote'
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
-import FavoriteIcon from '@material-ui/icons/Favorite'
 import MoreVertIcon from '@material-ui/icons/MoreVert'
-import NotificationImportantIcon from '@material-ui/icons/NotificationImportant'
-import TrendingUpIcon from '@material-ui/icons/TrendingUp'
-import clsx from 'clsx'
 import React from 'react'
+import Actions from '../ActionsPost'
+import Comments from '../Comments'
 
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({
@@ -46,22 +34,6 @@ const useStyles = makeStyles((theme: Theme) =>
         },
     })
 )
-
-const handleLike = () => {
-    alert('Gostei')
-}
-const handleImprove = () => {
-    alert('Melhorar')
-}
-const handleDenounce = () => {
-    alert('Denunciar')
-}
-const handleRead = () => {
-    alert('Ler')
-}
-const handleEdit = () => {
-    alert('Editar')
-}
 
 export default function Post() {
     const classes = useStyles()
@@ -99,72 +71,8 @@ export default function Post() {
                     peas along with the mussels, if you like.
                 </Typography>
             </CardContent>
-            <CardActions disableSpacing>
-                <IconButton aria-label="Gostei" onClick={handleLike}>
-                    <FavoriteIcon />
-                </IconButton>
-
-                <IconButton aria-label="Melhorar" onClick={handleImprove}>
-                    <TrendingUpIcon />
-                </IconButton>
-
-                <IconButton aria-label="Denunciar" onClick={handleDenounce}>
-                    <NotificationImportantIcon />
-                </IconButton>
-
-                <IconButton aria-label="Ler" onClick={handleRead}>
-                    <EventNoteIcon />
-                </IconButton>
-
-                <IconButton aria-label="Editar" onClick={handleEdit}>
-                    <EditIcon />
-                </IconButton>
-                <IconButton
-                    className={clsx(classes.expand, {
-                        [classes.expandOpen]: expanded,
-                    })}
-                    onClick={handleExpandClick}
-                    aria-expanded={expanded}
-                    aria-label="show more"
-                >
-                    <ExpandMoreIcon />
-                </IconButton>
-            </CardActions>
-            <Collapse in={expanded} timeout="auto" unmountOnExit>
-                <CardContent>
-                    <Typography paragraph>Method:</Typography>
-                    <Typography paragraph>
-                        Heat 1/2 cup of the broth in a pot until simmering, add
-                        saffron and set aside for 10 minutes.
-                    </Typography>
-                    <Typography paragraph>
-                        Heat oil in a (14- to 16-inch) paella pan or a large,
-                        deep skillet over medium-high heat. Add chicken, shrimp
-                        and chorizo, and cook, stirring occasionally until
-                        lightly browned, 6 to 8 minutes. Transfer shrimp to a
-                        large plate and set aside, leaving chicken and chorizo
-                        in the pan. Add pimentón, bay leaves, garlic, tomatoes,
-                        onion, salt and pepper, and cook, stirring often until
-                        thickened and fragrant, about 10 minutes. Add saffron
-                        broth and remaining 4 1/2 cups chicken broth; bring to a
-                        boil.
-                    </Typography>
-                    <Typography paragraph>
-                        Add rice and stir very gently to distribute. Top with
-                        artichokes and peppers, and cook without stirring, until
-                        most of the liquid is absorbed, 15 to 18 minutes. Reduce
-                        heat to medium-low, add reserved shrimp and mussels,
-                        tucking them down into the rice, and cook again without
-                        stirring, until mussels have opened and rice is just
-                        tender, 5 to 7 minutes more. (Discard any mussels that
-                        don’t open.)
-                    </Typography>
-                    <Typography>
-                        Set aside off of the heat to let rest for 10 minutes,
-                        and then serve.
-                    </Typography>
-                </CardContent>
-            </Collapse>
+            <Actions />
+            <Comments />
         </Card>
     )
 }
